LPD-89267 Inline class
